requirements.txt
absl-py==2.1.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
apache-libcloud==3.8.0
apache-tvm-ffi==0.1.6
argcomplete==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==25.3.0
audioread==3.0.1
av==15.1.0
azure-core==1.37.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
azure-storage-file-datalake==12.22.0
babel==2.16.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
bitsandbytes==0.47.0
blake3==1.0.5
bleach==6.2.0
blinker==1.9.0
boto3==1.42.16
botocore==1.42.16
Brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
circuitbreaker==2.1.3
click==8.2.1
cloudpickle==3.1.1
cmake==4.1.0
coloredlogs==15.0.1
colorful==0.5.7
comm==0.2.2
compressed-tensors==0.10.1
contourpy==1.3.1
coolname==2.2.0
cramjam==2.11.0
cryptography==45.0.7
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
cycler==0.12.1
databricks-sdk==0.76.0
datasets==3.6.0
debugpy==1.8.9
decorator==5.1.1
deepspeed==0.17.5
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
executing==2.1.0
fastapi==0.127.0
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastchat==0.1.0
fastjsonschema==2.20.0
fastrlock==0.8.3
ffmpy==0.6.1
filelock==3.19.1
fire==0.7.1
flashinfer-python==0.5.3
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.55.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2023.6.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.5.0
google-cloud-storage==2.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
gql==3.5.3
gradio_client==1.11.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.3.0
groovy==0.1.2
grpcio==1.68.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.33.5
humanfriendly==10.0
idna==3.10
ijson==3.4.0.post0
importlib_metadata==8.7.0
interegular==0.3.3
ipykernel==6.29.5
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json5==0.9.28
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.1
jupyterlab-language-pack-zh-CN==4.2.post3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.15.2
llguidance==0.7.30
llm-foundry==0.22.0
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.7.3
Mako==1.3.10
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.25.0
mdurl==0.1.2
mistral_common==1.8.5
mistune==3.0.2
mlflow==2.22.4
mlflow-skinny==2.22.4
model-hosting-container-standards==0.1.12
modelscope==1.29.2
mosaicml==0.32.1
mosaicml-cli==0.7.5
mosaicml-streaming==0.12.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cudnn-frontend==1.17.0
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-cutlass-dsl==4.3.4
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
oci==2.164.2
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.107.2
openai-harmony==0.0.8
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
orjson==3.11.3
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.15.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt-toolkit==3.0.36
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pynvml==11.5.3
pyOpenSSL==25.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-snappy==0.7.3
pytorch-ranger==0.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
questionary==2.0.1
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.6.4
rouge-chinese==1.0.3
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
ruff==0.13.0
s3transfer==0.16.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.16.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.37.1
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
shtab==1.7.2
six==1.16.0
slack_sdk==3.39.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==1.0.0
SQLAlchemy==2.0.45
sqlparse==0.5.5
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.50.0
supervisor==4.2.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.7.0
torch-optimizer==0.3.0
torchaudio==2.7.0
torchmetrics==1.7.4
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
trl==0.9.2
typeguard==4.4.4
typer==0.17.4
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
virtualenv==20.34.0
vllm==0.9.1
wandb==0.18.7
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.13
wrapt==1.17.3
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstd==1.5.7.2
